requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot